manimations/requirements.txt
gradio
openai
python-dotenv
manim